Let's get some OS X support going

+++ src/client/mac/testapp/Controller.m
@@ -53,7 +53,7 @@
 
 - (IBAction)forkTestOptions:(id)sender {
   NSInteger tag = [[sender selectedCell] tag];
-  NSLog(@"sender tag: %d", tag);
+  NSLog(@"sender tag: %ld", (long)tag);
   if (tag <= 2) {
     bpForkOption = tag;
   }
